Conditionally run validations rather than allowing blank
Allow_blank is confusing and contradictory here, since it's specifically
forbidden in a preceding validation. Only running the validation when
values are present is more communicative of what we want — namely, that
this is an edge case for something that should probably never happen
anyway (only by manually constructed parameters or data loss).