Add nonce value to stylesheet tags
This, it turns out was a thing Google changed sometime pretty close to
when this was reported. Hence, why I was pretty sure that I had gotten it
to work in the past.

Before approximately June, 2021, the GMaps JS would look at the nonce
value on it's own `<script>` tag and add that value to all the other
`<script>` and `<style>` tags that it inserts.  After then, they decided
that, because the nonce for JS and the nonce for CSS can be different,
it should use the nonce value from some other style tag for its own
`<style>` tags.

Closes #316, and for some reason appears to close #315