Don't make expectations on the downloads folder
All of these system tests are a nightmare. But, at some point these ones
made the transition from frequently failing to always failing. I believe
it has something to de with a change Chrome made to its headless mode.

And of course, these aren't even "downloads" in real life. They're
`Content-Disposition: inline`. So it's only the tests that ever saw them
as "downloads".