Fix require-order
Our prawn-rails-form opens up `PrawnRails::Document` to add onto it. But
since PrawnRails 1.5, that class isn't required initially, and our gem
doesn't specify that `PrawnRails::Document` inherits from
`Prawn::Document` which causes a `TypeError` when it eventually _is_
required.

This should be fixed in our gem, but this is a work-around