Upgrade prawn-rails to 1.6.0 (#733)
* Update prawn-rails to 1.6

* Fix require-order

Our prawn-rails-form opens up `PrawnRails::Document` to add onto it. But
since PrawnRails 1.5, that class isn't required initially, and our gem
doesn't specify that `PrawnRails::Document` inherits from
`Prawn::Document` which causes a `TypeError` when it eventually _is_
required.

This should be fixed in our gem, but this is a work-around

* Don't just `include` our module...everywhere

This defines the methods in our module on... I think every class?
Without it, `PrawnRailsForms.default_text_field_options` doesn't work
becuase they're not Module methods. This made it "work" by... defining
that method on every instance of `Object` (including every instance of
`Module`)